requirements.txt
python-dotenv
requests
geopandas